Moved eye-tracking tests outside backend
