+++ tests/eye_tracking/iris_tests/__init__.py
@@ -1,0 +1,2 @@
+from .process_eye_metrics import process_eye
+
